widget, widget/material: paint scrollbar as hovered while dragging it
Scrollbar.IndicatorHovered returns false while the scrollbar is being
dragged because Scrollbar.drag is grabbing the pointer while dragging.
Add a new method Dragging and use that. This would allow us to render
the indicator differently for hovering and dragging, but for now we just
draw it as hovered while it's being dragged.